backend/pyproject.toml
[project]
name = "vnstock-hub-backend"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt<4.0.0",
    "email-validator>=2.2.0",
    "fastapi>=0.128.0",
    "pip>=25.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "vnai>=2.3.9",
    "vnii>=0.1.9",
    "vnstock>=3.4.1",
    "pytz>=2025.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[[tool.uv.index]]
url = "https://vnstocks.com/api/simple"
